refactor!: move outputs and bindable interface to azfunc package
